test(tree): organize mock data in its own folder
